DINCAE.jl

DINCAE (Data-Interpolating Convolutional Auto-Encoder) is a neural network to reconstruct missing data in satellite observations. It can work with gridded data (DINCAE.reconstruct) or a clouds of points (DINCAE.reconstruct_points). In the later case, the data can be organized in e.g. tracks (or not).

The code is available at: https://github.com/gher-ulg/DINCAE.jl

The method is described in the following articles:

Barth, A., Alvera-Azcárate, A., Licer, M., and Beckers, J.-M.: DINCAE 1.0: a convolutional neural network with error estimates to reconstruct sea surface temperature satellite observations, Geosci. Model Dev., 13, 1609-1622, https://doi.org/10.5194/gmd-13-1609-2020, 2020.

Barth, A., Alvera-Azcárate, A., Troupin, C., and Beckers, J.-M.: DINCAE 2: multivariate convolutional neural network with error estimates to reconstruct sea surface temperature satellite and altimetry observations, Geosci. Model Dev. Discuss. [preprint], https://doi.org/10.5194/gmd-2021-353, 2022 (accepted).

The neural network will be trained on the GPU. Note convolutional neural networks can require a lot of GPU memory depending on the domain size. So far, only NVIDIA GPUs are supported by the neural network framework Knet.jl using in DINCAE (beside training on the CPU but which prohibitively slow).

User API

In most cases, a user does only need to interact with the function DINCAE.reconstruct or DINCAE.reconstruct_points.

DINCAE.reconstruct — Function

reconstruct(Atype,data_all,fnames_rec;...)

Train a neural network to reconstruct missing data using the training data set and periodically run the neural network on the test dataset. The data is assumed to be avaialable on a regular longitude/latitude grid (which is the case of L3 satellite data).

Mandatory parameters

Atype: array type to use

data_all: list of named tuples. Every tuple should have filename, and varname.

data_all[1] will be used for training (and perturbed to prevent overfitting). All others entries data_all[2:end] will be reconstructed using the training network at the epochs defined by save_epochs.

fnames_rec: vector of filenames corresponding to the entries data_all[2:end]

Optional parameters:

epochs: the number of epochs (default 1000)

batch_size: the size of a mini-batch (default 50)

enc_nfilter_internal: number of filter of the internal encoding layers (default [16,24,36,54])

skipconnections: list of layers with skip connections (default 2:(length(enc_nfilter_internal)+1))

clip_grad: maximum allowed gradient. Elements of the gradients larger than this values will be clipped (default 5.0).

regularization_L2_beta: Parameter for L2 reguliziation (default 0, i.e. no regularization)

save_epochs: list of epochs where the results should be saved (default 200:10:epochs)

is3D: Switch to apply 2D (is3D == false) or 3D (is3D == true) convolutions (default false)

upsampling_method: interpolation method during upsampling which can be either :nearest or :bilinear (default :nearest)

ntime_win: number of time instance within the time window. This number should be odd. (default 3)

learning_rate: intial learning rate of the ADAM optimizater (default 0.001)

learning_rate_decay_epoch: The exponential recay rate of the leaning rate. After learning_rate_decay_epoch the learning rate is halved. The learning rate is compute as learning_rate * 0.5^(epoch / learning_rate_decay_epoch). learning_rate_decay_epoch can be Inf for a constant learning rate (default)

min_std_err: minimum error standard deviation preving a division close to zero (default exp(-5) = 0.006737946999085467)

loss_weights_refine: The weigh of the individual refinement layers using in the cost function.

If loss_weights_refine has a single element, then there is no refinement. (default (1.,))

Note

Note that also the optional parameters should be to tuned for a particular application.

DINCAE.reconstruct_points — Function

DINCAE.reconstruct_points(T,Atype,filename,varname,grid,fnames_rec )

Mandatory parameters:

T: Float32 or Float64: float-type used by the neural network

Array{T} or KnetArray{T}: array-type used by the neural network.

filename: NetCDF file in the format described below.

varname: name of the primary variable in the NetCDF file

grid: tuple of ranges with the grid in the longitude and latitude direction e.g. (-180:1:180,-90:1:90).

fnames_rec: NetCDF file names of the reconstruction

Optional parameters:

jitter_std_pos: standard deviation of the noise to be added to the position of observation (default (5,5))

auxdata_files: gridded auxillary data file for a multivariate reconstruction. auxdata_files is an array of named tuples with the fields (filename, the file name of the NetCDF file, varname the NetCDF name of the primary variable and errvarname the NetCDF name of the expected standard deviation error). For example:

auxdata_files = [
  (filename = "big-sst-file.nc"),
   varname = "SST",
   errvarname = "SST_error")]

The data in the file should already be interpolated on the targed grid. The file structure of the NetCDF file is described in DINCAE.load_gridded_nc.

See DINCAE.reconstruct for other optional parameters.

An (minimal) example of the NetCDF file is:

netcdf all-sla.train {
dimensions:
	track = 9628 ;
	time = 7445528 ;
variables:
	int64 size(track) ;
		size:sample_dimension = "time" ;
	double dates(track) ;
		dates:units = "days since 1900-01-01 00:00:00" ;
	float sla(time) ;
	float lon(time) ;
	float lat(time) ;
	int64 id(time) ;
	double dtime(time) ;
		dtime:long_name = "time of measurement" ;
		dtime:units = "days since 1900-01-01 00:00:00" ;
}

The file should contain the variables lon (longitude), lat (latitude), dtime (time of measurement) and id (numeric identifier, only used by post processing scripts) and dates (time instance of the gridded field). The file should be in the contiguous ragged array representation as specified by the CF convention allowing to group data points into "features" (e.g. tracks for altimetry). Every feature can also contain a single data point.

Internal functions

DINCAE.load_gridded_nc — Function

lon,lat,time,data,missingmask,mask = load_gridded_nc(fname,varname; minfrac = 0.05)

Load the variable varname from the NetCDF file fname. The variable lon is the longitude in degrees east, lat is the latitude in degrees North, time is a numpy datetime vector, data_full is a 3-d array with the data, missingmask is a boolean mask where true means the data is missing and mask is a boolean mask where true means the data location is valid, e.g. sea points for sea surface temperature.

At the bare-minimum a NetCDF file should have the following variables and attributes:

netcdf file.nc {
dimensions:
        time = UNLIMITED ; // (5266 currently)
        lat = 112 ;
        lon = 112 ;
variables:
        double lon(lon) ;
        double lat(lat) ;
        double time(time) ;
                time:units = "days since 1900-01-01 00:00:00" ;
        int mask(lat, lon) ;
        float SST(time, lat, lon) ;
                SST:_FillValue = -9999.f ;
}

the first variable is for isoutput is none specified

DINCAE.NCData — Type

dd = NCData(lon,lat,time,data_full,missingmask;
            train = false,
            obs_err_std = 1.,
            jitter_std = 0.05)

Return a structure holding the data for training (train = true) or testing (train = false) the neural network. obs_err_std is the error standard deviation of the observations. The variable lon is the longitude in degrees east, lat is the latitude in degrees North, time is a numpy datetime vector, data_full is a 3-d array with the data and missingmask is a boolean mask where true means the data is missing. jitter_std is the standard deviation of the noise to be added to the data during training.